@@ -1,25 +1,25 @@
 name: augmented_reality_plugin_wikitude
 description: The Wikitude plugin to create augmented reality experiences in your apps.
-version: 9.8.0
+version: 9.9.0
 author: Wikitude GmbH <flutter@wikitude.com>
 homepage: https://www.wikitude.com/
 documentation: https://www.wikitude.com/external/doc/documentation/latest/flutter/
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   
-  json_annotation: ^3.0.0
+  json_annotation: ^4.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   
-  build_runner: ^1.7.1
-  json_serializable: ^3.2.3
+  build_runner: ^2.0.2
+  json_serializable: ^4.1.1
 
 flutter:
   plugin:
