@@ -1,19 +1,19 @@
 name: wikitude_flutter_app
 description: A new Flutter project.
 
-version: 9.8.0+1
+version: 9.9.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
-  location: ^2.3.5
-  path_provider: ^1.3.0
-  wakelock: ^0.1.3
+  cupertino_icons: ^1.0.3
+  location: ^4.1.1
+  path_provider: ^2.0.1
+  wakelock: ^0.5.1
 
   augmented_reality_plugin_wikitude:
     path: ./plugin/augmented_reality_plugin_wikitude/
@@ -22,8 +22,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^1.7.1
-  json_serializable: ^3.2.3
+  build_runner: ^2.0.2
+  json_serializable: ^4.1.1
 
 flutter:
 
