@@ -17,25 +17,24 @@ import 'package:augmented_reality_plugin_wikitude/wikitude_response.dart';
 import 'package:wikitude_flutter_app/poiDetails.dart';
 
 class ArViewState extends State<ArViewWidget> with WidgetsBindingObserver {
-  ArchitectWidget architectWidget;
+  late ArchitectWidget architectWidget;
   String wikitudeTrialLicenseKey = "rRcPFV/GWHOalFjHX9rP9TWGNRKVu8P4FSKvHtps1mo14SexXUmlVAebLNuKKr9OcOFD89RiMH03AY3eJL09d3Pbvb/V+AVYsQiBROkqqAhYe2lDojp++ZAPDx2RM9rJrD+1qYyUUbdUyKzIJXrU09u4tST9NdhER08njP2tMydTYWx0ZWRfX/p90uj/Yn9x/bcRTK6REaUg/GJT6uUKh7KfnXmxAtt0RI9WNjVPQFFjS1WFGtrRI43/VqyS0gnfsjmiov6fyrE+0aGBxJIzBNWupROE+AYw9LFkJ0gRN6KhsqvawIobvSPbVH+OaYanwnIV8q34LyTRujMzvJL+ke0hEfucf6eChYWe3O5kGCRD09oDnBzBLYnZotRjtuDb2eiHksj28kNuHJTlWItLA4A5Xjri7I1FmnCnTYezZfS2EHHazgOwfYAx+RMTSDXkdjrfradWo4kQFlERljYr1fXTh0T9s19r9FJTeao5/4UbUqcAW8mu71LoIQ5i2gJLDEp4d7xBEBaSznQ2TI4DSNW13lGlTXx8Ma47sFk4uxcxNy1S56RC1bPXA/iJGudxQrGMlhrYuwYcbEpKqEAqRB3xCZKV0M/69hlcZTreu3+1LbtYpLBFQ9GGMPC5FMjzVt29UFSVFyChB6PJlfVrpXbyvlq7ZFKWPc77HKIUyVhx5cSuI19pMxoTPiK5FfcuD7NeUJISK2loWzM/Cd5kvjqCZf0mGJ4zs9iwAQrkhpBGr07lwyAKJ0wH4ybZIdFXb69uZHnp9YnibYF6cuq5L+66lNPRicm1ojF46Sc6SkiVeZDfS6J1f2UOL1ymEMi3eH7pc8+AQ5JUn7XJWr8xIcYTlBa4HkJkRV7ire2Daij3cNywrcVv1GuReHLyW+UipWGPKrvY8IONHmkLEuAgdU9WupbmVdt24Cjn2s1n/ecIIKIVm9xgvdd5n4DHXKsOOWY03gp43g/5jgTJdl1PNwaVIvnwC1zMchAL5Ld49im8pcZbYiQC/MQqAdixxpORPZ0i6j0TM86K7P6DgSxmMNP/SG4vDx0m9mxvCIzvyevNl69Rc2yRToAwY1yGHMHyT2LwWr1NDhhW620ALR/u8gycvRhICYmISCwuCEBuSK+2UyKuKHk50gCr+xfLenxYshOJC+3dyGgBKXMkh/T8i0vKIBaKX5LcD0BY+msO4h/vrb4dMB61qzxCuJM8ax6O5tuQc4u5WOi/6XrAIRFTCqLMST8U6JKN689s70FJtvQYm0DpbPfYTOfeA53B5fphfsTMQqXFwKPhVLczCoWftmlLhHb/NcmNmCHnTp/Mm9yObyNsiG3oQ1Wbb1a9eMOcJ5y/Wvpi0RSYGwIfJcIIknvJIwPphZ3AJ3K9x/M89kct/J65XZMAMdnM1FbtLRpgKUVAUIUJ/E6V03QP/ElUHHukYjbXABWs/fJ/6uy9E4aXjbmzJQ6I9VKQ1uUsT2Oh8585HoXp6LLiFxADdRSIllJBtuMCmgfrd06qQ/q9wu8xFzvJYBeIT6xlCbsBXgdm";
   Sample sample;
   String loadPath = "";
   bool loadFailed = false;
 
-  ArViewState(Sample sample) {
-    this.sample = sample;
-    if(sample.path.contains("http://") || sample.path.contains("https://")) {
-      loadPath = sample.path;
+  ArViewState({required this.sample}) {
+    if(this.sample.path.contains("http://") || this.sample.path.contains("https://")) {
+      loadPath = this.sample.path;
     } else {
-      loadPath = "samples/" + sample.path;
+      loadPath = "samples/" + this.sample.path;
     }
   }
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
 
     architectWidget = new ArchitectWidget(
       onArchitectWidgetCreated: onArchitectWidgetCreated,
@@ -49,11 +48,9 @@ class ArViewState extends State<ArViewWidget> with WidgetsBindingObserver {
 
   @override
   void dispose() {
-    if (this.architectWidget != null) {
-      this.architectWidget.pause();
-      this.architectWidget.destroy();
-    }
-    WidgetsBinding.instance.removeObserver(this);
+    this.architectWidget.pause();
+    this.architectWidget.destroy();
+    WidgetsBinding.instance!.removeObserver(this);
 
     Wakelock.disable();
     super.dispose();
@@ -63,14 +60,10 @@ class ArViewState extends State<ArViewWidget> with WidgetsBindingObserver {
   void didChangeAppLifecycleState(AppLifecycleState state) {
     switch (state) {
       case AppLifecycleState.paused:
-        if (this.architectWidget != null) {
-          this.architectWidget.pause();
-        }
+        this.architectWidget.pause();
         break;
       case AppLifecycleState.resumed:
-        if (this.architectWidget != null) {
-          this.architectWidget.resume();
-        }
+        this.architectWidget.resume();
         break;
 
       default:
@@ -84,7 +77,12 @@ class ArViewState extends State<ArViewWidget> with WidgetsBindingObserver {
       body: WillPopScope(
         onWillPop: () async {
           if(defaultTargetPlatform == TargetPlatform.android && !loadFailed) {
-            return !(await this.architectWidget.canWebViewGoBack());
+            bool? canWebViewGoBack = await this.architectWidget.canWebViewGoBack();
+            if (canWebViewGoBack != null) {
+              return !canWebViewGoBack;
+            } else {
+              return true;
+            }
           } else {
             return true;
           }
@@ -100,13 +98,12 @@ class ArViewState extends State<ArViewWidget> with WidgetsBindingObserver {
     this.architectWidget.load(loadPath, onLoadSuccess, onLoadFailed);
     this.architectWidget.resume();
 
-    if(sample.requiredExtensions != null && sample.requiredExtensions.contains("application_model_pois")) {
-      ApplicationModelPois applicationModelPois = new ApplicationModelPois();
-      List<Poi> pois = await applicationModelPois.prepareApplicationDataModel();
+    if(sample.requiredExtensions.contains("application_model_pois")) {
+      List<Poi> pois = await ApplicationModelPois.prepareApplicationDataModel();
       this.architectWidget.callJavascript("World.loadPoisFromJsonData(" + jsonEncode(pois) + ");");
     }
     
-    if(sample.requiredExtensions != null && (sample.requiredExtensions.contains("screenshot") ||
+    if((sample.requiredExtensions.contains("screenshot") ||
         sample.requiredExtensions.contains("save_load_instant_target") ||
         sample.requiredExtensions.contains("native_detail"))) {
       this.architectWidget.setJSONObjectReceivedCallback(onJSONObjectReceived);
@@ -181,10 +178,10 @@ class ArViewWidget extends StatefulWidget {
   final Sample sample;
 
   ArViewWidget({
-    Key key,
-    @required this.sample,
+    Key? key,
+    required this.sample,
   });
 
   @override
-  ArViewState createState() => new ArViewState(sample);
+  ArViewState createState() => new ArViewState(sample: sample);
 }
