@@ -8,25 +8,25 @@ class Sample {
   List<String> requiredFeatures;
   StartupConfiguration startupConfiguration;
   
-  Sample({this.requiredExtensions, this.name, this.path, this.requiredFeatures, this.startupConfiguration});
+  Sample({required this.requiredExtensions, required this.name, required this.path, required this.requiredFeatures, required this.startupConfiguration});
 
   factory Sample.fromJson(Map<String, dynamic> jsonMap){
     var requiredExtensionsFromJson = jsonMap['required_extensions'];
-    List<String> requiredExtensionsList = new List();
+    List<String> requiredExtensionsList = [];
     if(requiredExtensionsFromJson != null) {
       requiredExtensionsList = new List<String>.from(requiredExtensionsFromJson);
     }
 
     var requiredFeaturesFromJson = jsonMap['requiredFeatures'];
-    List<String> requiredFeaturesList = new List();
+    List<String> requiredFeaturesList = [];
     if(requiredFeaturesFromJson != null) {
       requiredFeaturesList = new List<String>.from(requiredFeaturesFromJson);
     }
 
     var exampleStartupConfiguration = jsonMap['startupConfiguration'];
-    StartupConfiguration startupConfigurationItem;
+    StartupConfiguration startupConfigurationItem = StartupConfiguration();
     if(exampleStartupConfiguration != null) {
-      CameraPosition cameraPosition;
+      CameraPosition? cameraPosition;
       switch(exampleStartupConfiguration["camera_position"]) {
         case "back": cameraPosition = CameraPosition.BACK;
           break;
@@ -36,7 +36,7 @@ class Sample {
           break;
       }
 
-      CameraResolution cameraResolution;
+      CameraResolution? cameraResolution;
       switch(exampleStartupConfiguration["camera_resolution"]) {
         case "sd_640x480": cameraResolution = CameraResolution.SD_640x480;
           break;
@@ -48,7 +48,7 @@ class Sample {
           break;
       }
 
-      CameraFocusMode cameraFocusMode;
+      CameraFocusMode? cameraFocusMode;
       switch(exampleStartupConfiguration["camera_focus_mode"]) {
         case "once": cameraFocusMode = CameraFocusMode.ONCE;
           break;
