@@ -4,6 +4,12 @@
 
 #import "GeneratedPluginRegistrant.h"
 
+#if __has_include(<augmented_reality_plugin_wikitude/WikitudePlugin.h>)
+#import <augmented_reality_plugin_wikitude/WikitudePlugin.h>
+#else
+@import augmented_reality_plugin_wikitude;
+#endif
+
 #if __has_include(<location/LocationPlugin.h>)
 #import <location/LocationPlugin.h>
 #else
@@ -22,19 +28,13 @@
 @import wakelock;
 #endif
 
-#if __has_include(<augmented_reality_plugin_wikitude/WikitudePlugin.h>)
-#import <augmented_reality_plugin_wikitude/WikitudePlugin.h>
-#else
-@import augmented_reality_plugin_wikitude;
-#endif
-
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
+  [WikitudePlugin registerWithRegistrar:[registry registrarForPlugin:@"WikitudePlugin"]];
   [LocationPlugin registerWithRegistrar:[registry registrarForPlugin:@"LocationPlugin"]];
   [FLTPathProviderPlugin registerWithRegistrar:[registry registrarForPlugin:@"FLTPathProviderPlugin"]];
   [WakelockPlugin registerWithRegistrar:[registry registrarForPlugin:@"WakelockPlugin"]];
-  [WikitudePlugin registerWithRegistrar:[registry registrarForPlugin:@"WikitudePlugin"]];
 }
 
 @end
