@@ -4,9 +4,9 @@ enum CameraResolution { SD_640x480, HD_1280x720, FULL_HD_1920x1080, AUTO }
 
 class StartupConfiguration {
 
-  CameraPosition cameraPosition;
-  CameraFocusMode cameraFocusMode;
-  CameraResolution cameraResolution;
+  CameraPosition? cameraPosition = CameraPosition.BACK;
+  CameraFocusMode? cameraFocusMode = CameraFocusMode.CONTINUOUS;
+  CameraResolution? cameraResolution = CameraResolution.AUTO;
 
   StartupConfiguration({this.cameraPosition, this.cameraFocusMode, this.cameraResolution});
 }
